backend/requirements.txt
fastapi
uvicorn
scrapegraphai
python-dotenv